A lot of things happened...
Turns out cross-compiling is horrible, so I just threw it out all the C code and rewrote it in equally as horrible Java Code (I do plan on rewriting it in Kotlin eventually).

I also fixed an issues with internals not being resolved properly by commenting out all broken uses of the keyword internal in Kotlin until it's resolved.